requirements.txt
streamlit
pandas
numpy
investpy
yfinance
python-dateutil
matplotlib